Flexible z config (#847)
* Update OpenGLPanel.java

* turtle: use point and vector for p and n

* Update TurtleMove.java

* Change Turtle internals

instead of one history list with different types of moves, it is a list of layers each with a list of lines each with a lists of points.

* better error report in unit test

* compare without line endings

* restrict access to Turtle.strokeLayers

* Create ParallelLines.java

* bump version

donatello 1.8.6

* Update ThickenLinesByIntensity.java

* revert store raster

* document

* added flexible resolution

* fixed ProgramPanel parsing problems

Throwing exceptions in ancient code.  Wtf?

* added audio test file

* Update README.md

* 7.71.5

and bump versions

* added customizable gcode for pen up/down

* multipass UI uses slider for angle

* added sample rate to pulse

* 7.72.0

* prevent creation of redundant layers

when adding a layer, if it matches the previous layer, merge them together.

* 7.72.1